Make #each skip #shift-ed values again
Prior to 39da059c6d363136a80867a6cdb7f2ac42e48ecc, shifting a value from
a CSV, and then using an Enumerable method, would not include the
shifted value in the result of the Enumerable method.

This behavior made sense. The #shift methods on Array and Hash both
behave this same way, because they both remove an element from the
underlying data structure. It makes sense for CSV#shift to work like
this to, as it used to.

To fix this regression, there needs to be an internal position that is
advanced when calling #shift, which #each will then start from. By
re-implementing #each in terms of Enumerator#next, we get this for free.
This also allows #shift to be re-implemented in terms of #each.

It might seem that this would break using #each multiple times without
sending #rewind in between, but it turns out this has never been
supported anyway. I've added a test for this behaviour so as to
document it.