Fix line number in invalid byte sequence error
This didn't take into account that the sample could be composed of
multiple lines, and so the index of the malformed line in the sample
wasn't taken into account when calculating the line number.

I chose to just change the existing test for this to have the error on
line two, rather than writing a whole new test, because I think it's
much more likely that 1 will be returned incorrectly than 2, so this way
it's just an all around better test.