Fix case where an \r\n row separator will be split when reading a chunk
In this case, read one more character.

This is a suboptimal fix, as it doesn't fix handling of row
separators that aren't two characters and starting with \r.
A better fix would handle all multibyte row separators.
However, as \r\n is one of the most common row separators,
I think it's useful to merge this until a more generic
solution is developed.

Fixes [Bug #18245]